requirements.txt
opencv_python==4.12.0.88
Pillow==11.3.0
